Patch for windows build

+++ util/include/audiobuffer.h
@@ -87,7 +87,7 @@ public:
   }
 
   // This is the size in bytes (sizeof(fmt)*nSamples()!)
-  uint32_t size() const {
+  size_t size() const {
     return data_.size();
   }
